scripts/requirements.txt
polib
pyyaml
requests
